Updates the runboyrunband project to new format
Updates to move details into blog posts and add a basic overview
of the project into the project listing. This format will allow
better management of the information associated with each of the
projects. Project pages will include links to relevant services,
relevant blog posts, and a quick overview of the project.